+++ functions/ErrorMessage.fnc.php
@@ -148,7 +148,10 @@ function ErrorSendEmail( $error = [], $title = 'PHP Fatal error' )
 		return false;
 	}
 
-	$ip = issetVal( $_SERVER['HTTP_X_FORWARDED_FOR'], $_SERVER['REMOTE_ADDR'] );
+	$ip = ( isset( $_SERVER['HTTP_X_FORWARDED_FOR'] )
+		// Filter IP, HTTP_* headers can be forged.
+		&& filter_var( $_SERVER['HTTP_X_FORWARDED_FOR'], FILTER_VALIDATE_IP ) ?
+		$_SERVER['HTTP_X_FORWARDED_FOR'] : $_SERVER['REMOTE_ADDR'] );
 
 	$debug_backtrace = debug_backtrace();